update generate.py float -> int error

+++ generate.py
@@ -83,10 +83,8 @@ for _ in range(trainDelay):
     generated = np.append(generated, tmp, axis=0)
 
 if len(generated.shape) < 3:
-    generated = np.reshape(generated, (generated.shape[0], generated.shape[1]/2, 2))
+    generated = np.reshape(generated, (generated.shape[0], int(generated.shape[1]/2), 2))
 
 fnorm = utils.faceNormalizer()
-generated = fnorm.alignEyePointsV2(600*generated) / 600.0 
+generated = fnorm.alignEyePointsV2(600*generated) / 600.0
 utils.write_video_wpts_wsound(generated, sound, fs, output_path, 'PD_pts', [0, 1], [0, 1])
-
-
